refactor: 엔티티 plan 패키지 제거
