+++ src/main/java/com/yeoljeong/tripmate/application/dto/result/PlanUnitDetailResult.java
@@ -1,8 +1,7 @@
 package com.yeoljeong.tripmate.application.dto.result;
 
-import com.yeoljeong.tripmate.domain.model.plan.PlanParticipation;
-import com.yeoljeong.tripmate.domain.model.plan.PlanUnit;
-import com.yeoljeong.tripmate.presentation.dto.response.PlanUnitDetail;
+import com.yeoljeong.tripmate.domain.model.PlanParticipation;
+import com.yeoljeong.tripmate.domain.model.PlanUnit;
 import java.math.BigDecimal;
 import java.time.LocalTime;
 import java.util.List;
