+++ src/main/java/com/yeoljeong/tripmate/application/dto/result/PlanUnitParticipantResult.java
@@ -2,7 +2,7 @@ package com.yeoljeong.tripmate.application.dto.result;
 
 import com.yeoljeong.tripmate.domain.enums.ParticipationRole;
 import com.yeoljeong.tripmate.domain.enums.ParticipationStatus;
-import com.yeoljeong.tripmate.domain.model.plan.PlanParticipation;
+import com.yeoljeong.tripmate.domain.model.PlanParticipation;
 import java.util.List;
 import java.util.UUID;
