+++ src/main/java/com/yeoljeong/tripmate/application/service/command/PlanCommandService.java
@@ -13,15 +13,15 @@ import com.yeoljeong.tripmate.domain.enums.ParticipationRole;
 import com.yeoljeong.tripmate.domain.enums.ParticipationStatus;
 import com.yeoljeong.tripmate.domain.events.PlanEvents;
 import com.yeoljeong.tripmate.domain.exception.PlanErrorCode;
-import com.yeoljeong.tripmate.domain.model.plan.PlanProductSnapshot;
+import com.yeoljeong.tripmate.domain.model.PlanProductSnapshot;
 import com.yeoljeong.tripmate.domain.provider.ProductData;
 import com.yeoljeong.tripmate.domain.provider.ProductProvider;
 import com.yeoljeong.tripmate.domain.repository.PlanParticipationRepository;
 import com.yeoljeong.tripmate.domain.repository.PlanProductSnapshotRepository;
 import com.yeoljeong.tripmate.domain.repository.PlanRepository;
-import com.yeoljeong.tripmate.domain.model.plan.Plan;
-import com.yeoljeong.tripmate.domain.model.plan.PlanParticipation;
-import com.yeoljeong.tripmate.domain.model.plan.PlanUnit;
+import com.yeoljeong.tripmate.domain.model.Plan;
+import com.yeoljeong.tripmate.domain.model.PlanParticipation;
+import com.yeoljeong.tripmate.domain.model.PlanUnit;
 import com.yeoljeong.tripmate.domain.repository.PlanUnitRepository;
 import com.yeoljeong.tripmate.event.EventUtils;
 import com.yeoljeong.tripmate.exception.BusinessException;
