+++ src/main/java/com/yeoljeong/tripmate/domain/repository/PlanProductSnapshotRepository.java
@@ -1,6 +1,6 @@
 package com.yeoljeong.tripmate.domain.repository;
 
-import com.yeoljeong.tripmate.domain.model.plan.PlanProductSnapshot;
+import com.yeoljeong.tripmate.domain.model.PlanProductSnapshot;
 
 public interface PlanProductSnapshotRepository {
