+++ src/main/java/com/yeoljeong/tripmate/infrastructer/persistence/jpa/PlanJpaRepository.java
@@ -1,6 +1,6 @@
 package com.yeoljeong.tripmate.infrastructer.persistence.jpa;
 
-import com.yeoljeong.tripmate.domain.model.plan.Plan;
+import com.yeoljeong.tripmate.domain.model.Plan;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
