+++ src/main/java/com/yeoljeong/tripmate/infrastructer/persistence/repository/PlanRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.yeoljeong.tripmate.infrastructer.persistence.repository;
 
-import com.yeoljeong.tripmate.domain.model.plan.Plan;
+import com.yeoljeong.tripmate.domain.model.Plan;
 import com.yeoljeong.tripmate.domain.repository.PlanRepository;
 import com.yeoljeong.tripmate.infrastructer.persistence.jpa.PlanJpaRepository;
 import java.util.Optional;
